+++ src/biodwh2-pathwaycommons/src/main/java/de/unibi/agbi/biodwh2/pathwaycommons/etl/PathwayCommonsGraphExporter.java
@@ -1,0 +1,23 @@
+package de.unibi.agbi.biodwh2.pathwaycommons.etl;
+
+import de.unibi.agbi.biodwh2.core.Workspace;
+import de.unibi.agbi.biodwh2.core.etl.GraphExporter;
+import de.unibi.agbi.biodwh2.core.exceptions.ExporterException;
+import de.unibi.agbi.biodwh2.core.model.graph.Graph;
+import de.unibi.agbi.biodwh2.pathwaycommons.PathwayCommonsDataSource;
+
+public class PathwayCommonsGraphExporter extends GraphExporter<PathwayCommonsDataSource> {
+    public PathwayCommonsGraphExporter(final PathwayCommonsDataSource dataSource) {
+        super(dataSource);
+    }
+
+    @Override
+    public long getExportVersion() {
+        return 1;
+    }
+
+    @Override
+    protected boolean exportGraph(final Workspace workspace, final Graph graph) throws ExporterException {
+        return false;
+    }
+}