+++ src/biodwh2-pathwaycommons/src/main/java/de/unibi/agbi/biodwh2/pathwaycommons/etl/PathwayCommonsParser.java
@@ -1,0 +1,17 @@
+package de.unibi.agbi.biodwh2.pathwaycommons.etl;
+
+import de.unibi.agbi.biodwh2.core.Workspace;
+import de.unibi.agbi.biodwh2.core.etl.Parser;
+import de.unibi.agbi.biodwh2.core.exceptions.ParserException;
+import de.unibi.agbi.biodwh2.pathwaycommons.PathwayCommonsDataSource;
+
+public class PathwayCommonsParser extends Parser<PathwayCommonsDataSource> {
+    public PathwayCommonsParser(final PathwayCommonsDataSource dataSource) {
+        super(dataSource);
+    }
+
+    @Override
+    public boolean parse(final Workspace workspace) throws ParserException {
+        return true;
+    }
+}